ml/requirements.txt
tensorflow==2.15.0
keras==2.15.0
flask
flask-cors
numpy
pillow
joblib
scikit-learn
